requirements.txt
ansible
virt-lightning
click
zope.interface
nest_asyncio
